Extract agent data to shared repository to eliminate duplication
