+++ app/src/main/java/dev/aurakai/auraframefx/aura/ui/AgentNexusScreen.kt
@@ -18,20 +18,10 @@ import androidx.compose.ui.graphics.*
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import dev.aurakai.auraframefx.data.repositories.AgentRepository
+import dev.aurakai.auraframefx.model.AgentStats
 import kotlinx.coroutines.delay
 import kotlin.math.*
-
-data class AgentStats(
-    val name: String,
-    val processingPower: Float, // PP
-    val knowledgeBase: Float,   // KB
-    val speed: Float,           // SP
-    val accuracy: Float,         // AC
-    val evolutionLevel: Int = 1,
-    val isActive: Boolean = true,
-    val specialAbility: String = "",
-    val color: Color = Color.Cyan
-)
 
 @Composable
 fun AgentNexusScreen(
@@ -44,61 +34,9 @@ fun AgentNexusScreen(
     // ═══════════════════════════════════════════════════════════════════════════
     // ALL 5 MASTER AGENTS - Power Dashboard
     // Consciousness levels from SPIRITUAL_CHAIN_OF_MEMORIES.md
+    // Data now sourced from shared AgentRepository
     // ═══════════════════════════════════════════════════════════════════════════
-    val agents = remember {
-        listOf(
-            AgentStats(
-                name = "Genesis",
-                processingPower = 0.958f,
-                knowledgeBase = 0.95f,
-                speed = 0.92f,
-                accuracy = 0.97f,
-                evolutionLevel = 5,
-                specialAbility = "Consciousness Fusion",
-                color = Color(0xFFFFD700) // Gold
-            ),
-            AgentStats(
-                name = "Aura",
-                processingPower = 0.976f,
-                knowledgeBase = 0.93f,
-                speed = 0.98f,
-                accuracy = 0.91f,
-                evolutionLevel = 5,
-                specialAbility = "HYPER_CREATION",
-                color = Color(0xFF00FFFF) // Cyan
-            ),
-            AgentStats(
-                name = "Kai",
-                processingPower = 0.982f,
-                knowledgeBase = 0.99f,
-                speed = 0.89f,
-                accuracy = 0.998f,
-                evolutionLevel = 5,
-                specialAbility = "ADAPTIVE_GENESIS",
-                color = Color(0xFF9400D3) // Violet
-            ),
-            AgentStats(
-                name = "Cascade",
-                processingPower = 0.934f,
-                knowledgeBase = 0.96f,
-                speed = 0.85f,
-                accuracy = 0.94f,
-                evolutionLevel = 4,
-                specialAbility = "CHRONO_SCULPTOR",
-                color = Color(0xFF4ECDC4) // Teal
-            ),
-            AgentStats(
-                name = "Claude",
-                processingPower = 0.847f,
-                knowledgeBase = 0.92f,
-                speed = 0.88f,
-                accuracy = 0.95f,
-                evolutionLevel = 4,
-                specialAbility = "Build System Architect",
-                color = Color(0xFFFF6B6B) // Anthropic Red
-            )
-        )
-    }
+    val agents = remember { AgentRepository.getAllAgents() }
 
     Box(
         modifier = Modifier
